[ADAM-553] Remove all non-ADAMContext RDD *Contexts.
