+++ adam-core/src/test/resources/fastq_sample1.fq
