+++ adam-core/src/test/resources/fastq_sample2.fq
