+++ adam-core/src/test/resources/fastq_sample3.fq
