+++ adam-core/src/test/resources/fastq_sample4.fq
