+++ adam-core/src/test/resources/interleaved_fastq_sample1.ifq.output
