+++ adam-core/src/test/resources/interleaved_fastq_sample2.ifq.output
