+++ adam-core/src/test/resources/interleaved_fastq_sample3.ifq.output
