+++ adam-core/src/test/resources/interleaved_fastq_sample4.ifq.output
