+++ adam-core/src/test/resources/interleaved_multiline_fastq.ifq.output
